+++ src/components/Nav/Nav.scss
@@ -37,11 +37,14 @@
           width: 200px;
           height: 30px;
           padding-left: 45px;
-          color: #cacaca;
+          color: #7c7c7c;
           background-color: #efefef;
           outline: 0;
           border: 0;
           border-radius: 3px;
+        }
+        .nav-search-input::placeholder {
+          color: #cacaca;
         }
       }
       .nav-icon {
