+++ src/pages/Hyunjung/Main/FeedsArticle/FeedsArticle.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const FeedsArticle = ({ feedsArr }) => {
+  return (
+    <div className="feeds-article">
+      <span className="feeds-article-id">
+        <strong> {feedsArr.articleName} </strong>
+      </span>
+      <span className="feeds-article-content"> {feedsArr.articleContent} </span>
+      <span className="feeds-article-over">더 보기</span>
+    </div>
+  );
+};
+
+export default FeedsArticle;
